rename #64: 상세 페이지 관련 api 요청과 쿼리들 분리 및 기존 구조 변경
- itemDetails 관련 axios 요청, 리액트 쿼리, msw 핸들러들을 별도 파일로 분리
- queryKey 별도 파일로 관리
- api 폴더 내 fetch 함수들 fetchers 폴더로 이동
