+++ src/api/fetchers/itemFetcher.ts
@@ -1,5 +1,5 @@
-import { BASE_URL, fetcher } from './axios';
-import { API_ENDPOINT } from './endPoint';
+import { BASE_URL, fetcher } from '../axios';
+import { API_ENDPOINT } from '../endPoint';
 
 export const getItems = async ({
   pageParam: cursor,
@@ -15,6 +15,12 @@ export const getItems = async ({
     url.searchParams.append('categoryId', String(categoryId));
 
   const res = await fetcher.get(url.toString());
+
+  return res.data;
+};
+
+export const deleteItem = async (itemId: number) => {
+  const res = await fetcher.delete(`${API_ENDPOINT.ITEMS}/${itemId}`);
 
   return res.data;
 };
@@ -64,40 +70,6 @@ export const getFavoritesCategories = async () => {
 
 export const getCategories = async () => {
   const res = await fetcher.get(API_ENDPOINT.CATEGORIES);
-
-  return res.data;
-};
-
-export const getItemDetails = async (itemId: number) => {
-  const res = await fetcher.get(`${API_ENDPOINT.ITEMS}/${itemId}`);
-
-  return res.data;
-};
-
-export const patchFavorite = async ({
-  itemId,
-  isFavorite,
-}: {
-  itemId: number;
-  isFavorite: boolean;
-}) => {
-  const res = await fetcher.patch(`${API_ENDPOINT.ITEMS}/${itemId}/favorites`, {
-    isFavorite,
-  });
-
-  return res.data;
-};
-
-export const patchStatus = async ({
-  itemId,
-  statusName,
-}: {
-  itemId: number;
-  statusName: '판매중' | '예약중' | '판매완료';
-}) => {
-  const res = await fetcher.patch(`${API_ENDPOINT.ITEMS}/${itemId}/status`, {
-    statusName,
-  });
 
   return res.data;
 };
