+++ src/api/queries/queryKeys.ts
@@ -1,0 +1,8 @@
+export const QUERY_KEY = {
+  ITEMS: 'items',
+  ITEM_DETAILS: 'itemDetails',
+  SALES_LIST: 'salesList',
+  FAVORITES: 'favorites',
+  USER_LOCATION: 'userLocations',
+  LOCATION: 'locations',
+};
