+++ src/mocks/handlers.ts
@@ -1,4 +1,5 @@
 import { authHandlers } from './authHandlers';
+import { itemDetailsHandlers } from './itemDetailsHandlers';
 import { itemsHandlers } from './itemsHandlers';
 import { locationHandlers } from './locationHandlers';
 
@@ -6,4 +7,5 @@ export const handlers = [
   ...authHandlers,
   ...locationHandlers,
   ...itemsHandlers,
+  ...itemDetailsHandlers,
 ];
