+++ src/page/auth/SignUpPanel.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, useEffect, useRef, useState } from 'react';
 import { styled } from 'styled-components';
-import { singup } from '../../api/authFetcher';
+import { singup } from '../../api/fetchers/authFetcher';
 import { useResetLocationResult } from '../../api/queries/useLocationQuery';
 import { Header } from '../../components/Header';
 import { Button } from '../../components/button/Button';
